Added collections templating field labels to all Data Editor tools This includes: - Quick search from the collection profile page - Create new records using image - Skeletal data entry - Determination History tab of the occurrence editor - Add batch determinations - Specimen label printing - Annotation label printing - Outgoing loan specimens
Previously, these field labels were only used in the occurrence editor, but if set, it would make sense to have them the same throughout at least the Data Editor tools for consistency.
Where these fields are defined in language files, the collections template takes priority, falling back to the language file definition if the collection template is not set.

One exception to this is the list of fields in the Custom Field dropdown in the occurrence editor search query form. Since this pulls field names from elsewhere, I've set it up to look for a field label named [FIELDNAME]LABEL. Since some of these fields are not present in the collection editor template, they aren't customizable in this way.