Adds a binomial name field to label maker GUI, allows all parts of name to be styled
Previously, the visual interface had fields for scientific name, infraspecific epithet, taxon rank, scientific name authorship, and parent authorship, and would expose parent authorship to the labelmaker when the `Print species authors for infraspecific taxa` option was checked. However, this did not work in a case where both the parent authorship and the infrataxon authorship should be displayed, for example: `**Symphoricarpos albus** (L.) S.F. Blake var. **laevigatus** (Fernald) S.F. Blake`. In this example, since the scientific name was `**Symphoricarpos albus** var. **laevigatus**`, there was no way to insert the parent author correctly. This format is a common one for plant labels.

This fix enables a new field `Binomial Name` to the label maker GUI that is just the genus and species, without the infrataxon (represented internally as speciesname). That allows each aspect of the name to be included in the proper order, and to be themed separately, e.g., the binomial name and the infraspecific epithet are typically italicized (and sometimes bolded), while the taxon rank and the authorships are not.

One issue is that for taxonomic names not in the thesaurus, there are no fields by default for binomial name, taxon rank, and infraspecific epithet. The code detects this case and attemps to parse out those sections from the scientific name in the occurrence record.
Furthermore, for autonyms (where the infraspecific epithet matches the species epithet), the infaspecific author is typically not included, because it is the same as the species author by definition. The code detects this as well, and removes the infraspecific author if it is specified.