Adds util function to make use of mysqli_execute_query for older versions of php
# Summary
`mysqli_execute_query` or `$mysqli->execute_query` function is an 8.2
only function provided by mysqli which could break portals that cannot
upgrade to 8.2 from 8 or are running 7.3. Solution is to make a util
function with the same api as mysqli_execute_query but will use the php
version to decide wether to use mysqli_execute_query or normal statement
binding.