Closes #2119 Fixes Media Form History
# Issue #2119

# Summary
After search variables that were previously selected were not being
overidden by changes in the form. This was do to altering the form
action in order to preserve the url after a post. Fix was to load this
with js so that the history is correct but does not pollute the next
request.