Add Syncing Behavior Between Verbatim Coordinates and Latitude and Longitude
# Issue Closes #666

# Summary
Users desired for coordinates to not be truncated on input and due to
the limitations of storing decimal numbers in the database we settle for
automatically syncing them to the database.

There are other ways to tackle this issue that revolve around changing
the latitude and longitude fields into strings. However, that change
will likely have rippling effects on other parts of the code assuming
this is a number so futher dicussion will needed if these changes do not
suffice.