Closes #2365 Fixes Issue with Material Sample Form Names
# Issue #2365

# Summary
At some point form names got swapped to 'ms_[field_name]' so the insert
was not recoginzing the input. Update the form to check for both
'ms_[field_name]' and '[field_name]' variants because this function is
also used in the import which is where I imagine this got messed up at.