Closes #2448
# Summary
Adds Mannual Entry form that was previously in the polycoord tool. This
is to fix loss of functionality introduce when consolidating the
coordinate helper tools.

Some additional changes were made to mapAidUtils to not provide a
default polygon input id if not in polygon mode and to only reset forms
that are capture in the scope of the tool ie if strictly in polygon mode
then bounds set by a rectangle tool will not be overritten.