Adds Support for Custom GeoJSON Overlays
# Issue Closes #2516

# Summary
Allows users to configure GeoJSON Overlays that are toggable on the map
search page.

Currently Supports custom popup templates although support is minimal
and this feature may not be desired.

All config is currently managed in symbini and with root server access
to upload the desired GeoJSON files. See template for configuration
instructions