Closes #2513 QOL Checklist Polygon Input
# Issue closes #2513
# Summary
Enhances the checklist footprint form so that if changes are made to the
footprint the checkbox for searching it is automatically checked. Also
if the user uses the helper tool it will also be automatcially checked.
If the json value in the text area is not valid json it will display a
message stating so and not check the search box.