# Closes 2541 Fixes issue with sortSequence or sortsequence not populating when adding media
# Issue Closes #2541
# Summary
Because of the history of symbiota their is a schism in schema values
with sortsequence being used in older portals and sortSequence being
used in others. Some of that general confusion has bled into the logic.

To fix that some null coalencence to the correct value is needed. The
key is being stored as the lower case version `sortsequence` but because
of case insensitivity this will not be an issue at the insert level