Closes #2505 Change web url generated tag
# Issue #2505
# Summary
During refactor of image tooling to a more general media tooling there
was some confusion in the naming standards for images. Depending on the
resolution of the uploaded image the derivative would be either smaller
or bigger. The name change will better indicate that the url file is in
a certain target resolution and the original should be kept as is
resizing is needed again.