Bugfix: zeros are now shown when reviewing uploaded data
There was a bug where if a field was zero, it would be shown as a blank field in the upload reviewer. Additionally, if all records were null or zero for a field, then the field would not be shown. This caused issues with variables such as elevation that could have values of zero.