Adds copyright info and rights URL for individual media on public record displays
Previously, if the media copyright field was set, the rights for the record itself were set to match the media. However, this is not great form, as both the data and the media can and often do have different rights. For example, many collections license data as CC-BY-NC, and images as CC-BY-NC-SA. Additionally, either the data or the media may have more restrictive licences than the other component. For example, the data may be licensed under a creative commons license, while the media may be public domain (thus making the entire record public domain under the current paradigm). Or the media  could be copyrighted by a different individual and used with permission, while the data copyright is  more permissive.

A better way of handling this is to allow both sets of rights to be displayed, so this adds the image copyright (if set) to be displayed below the media.