Add host check to rss feed publisher
# Issue (None)

# Summary
Optimizes rss feeds to get local file information if they are stored at
the same host. Previously files store on the same host would make http
requests to themselves which was causing the php-fpm process to hang.
The exact for search for why fpm is having a problem is still ongoing.