Deprecate OccurrenceSkeletal in favor of using OccurrenceEditorManager
# Issue
No marked issue but related to #2826

# Summary
The `collections/editor/skeletalsubmit.php` file needs to add and update
records leveraging the already in existence class
OccurrenceEditorManager can cover the updating and deletion with the
missing functons `catalogNumberExists` and `getLanguageArr` which are
for the most part directly copied.

## Reasoning
Consolidating logic to as close to one place as possible makes it easier
to change and manage.