Add Render Sciname function for formatting taxonomy
# Issue #https://github.com/BioKIC/symbiota-docs/issues/679

# Summary
There are other places that do a similar job, however none of them just
take a string and render it without much of a fuss.

This function is not perfect but it will beat just rendering everything
in italics which is how this is mostly handled. Notably it doesn't
handle tradenames as I could not find much info of parsing indicators
for tradenames. If they are always tied to genus + species it would be
possible to infer them.