Fix register function to only use Bcrypt if enabled
# Summary
Account creation was using hashing regardless of wether USE_BCRYPT flag
was enabled. This was causing a problem where non bcyrpt enabled portals
could not create users without proper schema updates. It also prevent
the login of a user after account creation because the hash did not
match what was expected when using the authenticate function.