Adjust Material Sample language be be same as elsewhere
# Summary
`materialSampleVars.en.php` was not assigning to the $LANG array which
is not consitent with how we handle language loading. This difference
was breaking the lanugage loader and will break in Symbiota-Laravel
instances because it is introducing a previously unknown global variable

It also did not have translations for the fields which might need to
spun into a sperate issue.