Deprecate and Remove OccurrenceEditorImages
# Summary
PR to cleanup class that is no longer being used since being replaced by
the Media class.

This pr removes the remaining references to this class and shows the
resons for doing so.

## collections/editor/imageoccursubmit.php
There is no OccurrenceEditorImages specific methods being used so the
parent class is replacing it.

## Webservices/imageuploader.php
This page looks like it was work in progess that no longer works given
the main class it is trying to call doesn't exist and
`OccurrenceEditorImages` is instantiated but not called anywhere making
it safe to remove here.