Remove Remote DWC-Archive filesize support
# Summary
Will make external dwc archives not try to make header calls to get
filesizes of archives.

As far as I can tell this already cannot happen based on the fact that
the dwc publisher writes to the eml file using the results of
`GeneralUtil::getDomain` which will always be what $SERVER_HOST is
meaning their is only self hosted files. If we wish to support this the
files sizes should be stored in the eml file and read out instead of
fetching them dynamically.

Previous failure of this function would lock pages now the failure case
is a ? mark on the size which I see as an improvment.